functions/import-products/requirements.txt
il-supermarket-scraper>=0.18
il-supermarket-parser>=0.2
google-cloud-firestore>=2.16
google-cloud-logging>=3.10
